+++ client/src/components/Box.tsx
@@ -5,28 +5,16 @@ import { useParams, useHistory } from 'react-router-dom'
 import BoxContents from './BoxContents'
 import PasswordInput from './PasswordInput'
 import Screen from '../style/Screen'
+import { ErrorResponse, GetBoxesResponse } from '../types/Response'
 
 interface BoxParams {
   boxid: string
-}
-
-export interface ErrorResponse {
-  title: string
-  status: number
 }
 
 export interface Item {
   id: string
   name: string
   expirationDate: Date
-}
-
-interface GetBoxesResponse {
-  id: string
-  name: string
-  passwordRequired: boolean
-  items: Item[]
-  updatedAt: Date
 }
 
 const Box: React.FC = () => {
@@ -38,7 +26,6 @@ const Box: React.FC = () => {
   const [boxName, setboxName] = useState('')
   const [items, setItems] = useState<Item[]>([])
 
-
   useEffect((): void => {
     setBoxInfo()
   }, [])
@@ -49,7 +36,7 @@ const Box: React.FC = () => {
 
   const setBoxInfo = async () => {
     const boxInfo = await getBoxes()
-    if (boxInfo === undefined) {
+    if (boxInfo instanceof Error) {
       history.push('/')
       return
     }
@@ -73,11 +60,20 @@ const Box: React.FC = () => {
     setItems(boxInfo.items)
   }
 
-  const getBoxes = async (): Promise<GetBoxesResponse | ErrorResponse | undefined> => (
-    await axios.get<GetBoxesResponse>(`${process.env.REACT_APP_API_SERVER}/boxes/${boxid}`)
-      .then(res => (res.data))
-      .catch((err: AxiosError<ErrorResponse>) => (err.response?.data))
-  )
+  const getBoxes = async (): Promise<GetBoxesResponse | ErrorResponse | Error> => {
+    try {
+      const res = await axios.get<GetBoxesResponse>(`${process.env.REACT_APP_API_SERVER}/boxes/${boxid}`)
+      return res.data
+    } catch(err) {
+      if (err.response as AxiosError<ErrorResponse>) {
+        return err.response.data
+      }
+      if (err instanceof Error) {
+        return err
+      }
+      return new Error
+    }
+  }
 
   return (
     <BoxBody>
