+++ client/src/components/Card.tsx
@@ -1,34 +1,48 @@
 import React from 'react'
+import axios, { AxiosError } from 'axios'
 import styled from 'styled-components'
 import SaveFile from './Savefile'//デバッグ用
-import { Item, ErrorResponse } from './Box'
-import axios, { AxiosError } from 'axios'
-
-
+import { Item } from './Box'
+import { ErrorResponse, ItemResponse } from '../types/Response'
 
 interface ItemProps {
   boxid: string
   itemid: string
 }
 
-export interface ItemResponse {
-  id: string
-  name: string
-}
-
 const Card: React.FC<ItemProps> = ({ boxid, itemid }: ItemProps) => {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const getItem = async (): Promise<Item | ErrorResponse | undefined> => (
-    await axios.get<Item>(`${process.env.REACT_APP_API_SERVER}/boxes/${boxid}/${itemid}`)
-      .then(res => (res.data))
-      .catch((err: AxiosError<ErrorResponse>) => (err.response?.data))
-  )
+  const getItem = async (): Promise<Item | ErrorResponse | Error> => {
+    try {
+      const res = await axios.get<Item>(`${process.env.REACT_APP_API_SERVER}/boxes/${boxid}/${itemid}`)
+      return res.data
+    } catch(err) {
+      if (err.response as AxiosError<ErrorResponse>) {
+        return err.response.data
+      }
+      if (err instanceof Error) {
+        return err
+      }
+      return new Error
+    }
+  }
+
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const deleteItem = async (): Promise<ItemResponse | ErrorResponse | undefined> => (
-    await axios.delete<ItemResponse>(`${process.env.REACT_APP_API_SERVER}/boxes/${boxid}/${itemid}`)
-      .then(res => (res.data))
-      .catch((err: AxiosError<ErrorResponse>) => (err.response?.data))
-  )
+  const deleteItem = async (): Promise<ItemResponse | ErrorResponse | Error> => {
+    try {
+      const res = await axios.delete<ItemResponse>(`${process.env.REACT_APP_API_SERVER}/boxes/${boxid}/${itemid}`)
+      return res.data
+    } catch(err) {
+      if (err.response as AxiosError<ErrorResponse>) {
+        return err.response.data
+      }
+      if (err instanceof Error) {
+        return err
+      }
+      return new Error
+    }
+  }
+
   const saveDownloadFile = () => {
     SaveFile(new File(["Hello,World!"], "template.txt"))
   }
@@ -57,7 +71,6 @@ const CardDefault = styled.button`
     top: 0;
     left: 0;
   }
-
 `
 
 export default Card
