+++ client/src/types/Response.d.ts
@@ -1,0 +1,17 @@
+export interface ErrorResponse {
+  title: string
+  status: number
+}
+
+export interface GetBoxesResponse {
+  id: string
+  name: string
+  passwordRequired: boolean
+  items: Item[]
+  updatedAt: Date
+}
+
+export interface ItemResponse {
+  id: string
+  name: string
+}
